Update to BSM-Public
This update to BSM-Public includes revised model files, a revised set of default input data spreadsheets, and a revised searchable definition spreadsheet. Model changes are summarized in the file 2019-08 BSM-public high level changes.docx. A detailed list of changes to the model can be found in the file BSM-Public Difference 2019.xlsx